first commit to github

+++ fixcode.sh
@@ -16,7 +16,6 @@ if [ `echo ${INPUT: -1}` = "/" ]; then
 fi
 
 CURRENT_DIR=$(cd $(dirname $0); pwd)
-CONF_DIR=$CURRENT_DIR/config_bak/${INPUT}
 
 cd $CURRENT_DIR/$INPUT
 
@@ -25,89 +24,22 @@ if [ ! -f ./scripts/feeds ];then
     exit 1
 fi
 
-echo 'Reset Openwrt Source...'
+if [ -d .git ]; then
+    echo 'Reset Openwrt Source...'
+    git fetch --all && git clean -df && git reset --hard
 
-git fetch --all && git clean -df && git reset --hard
-
-echo -e '***Done***\n'
-
-read -r -p "Add GFW Feeds ? [Y/n] " input
-
-case $input in
-    [yY][eE][sS]|[yY])
-        echo 'Modify Default Feeds...'
-#       echo "src-git helloworld https://github.com/fw876/helloworld.git" >> "feeds.conf.default"
-#       echo "src-git openclash https://github.com/vernesong/OpenClash.git" >> "feeds.conf.default"
-        echo "src-git passwall_packages https://github.com/xiaorouji/openwrt-passwall.git;packages" >> "feeds.conf.default"
-        echo "src-git passwall_luci https://github.com/xiaorouji/openwrt-passwall.git;luci" >> "feeds.conf.default"
-        echo -e '***Done***\n'
-        ;;
-
-    [nN][oO]|[nN])
-        ;;
-
-    *)
-        echo "Invalid input..."
+    if [ $? -ne 0 ]; then
         exit 1
+    fi
+    echo -e '***Done***\n'
+
+    read -r -p "Modify Default IP ? [Y/n] " input
+    case $input in
+        [yY][eE][sS]|[yY])
+            echo 'Modify Default IP...'
+            sed -i "s/192.168.1.1/10.0.0.2/;s/192.168/10.0/" package/base-files/files/bin/config_generate
+
+            echo -e '***Done***\n'
         ;;
-esac
-
-read -r -p "Do you have custom config files ? [Y/n] " input
-
-case $input in
-    [yY][eE][sS]|[yY])
-        echo 'Add custom config files...'
-        ln -sf $CONF_DIR/files
-
-        echo -e '***Done***\n'
-       ;;
-
-    [nN][oO]|[nN])
-        if [ -L files ]; then
-            rm files
-        fi
-
-        read -r -p "Modify Default IP ? [Y/n] " input
-        case $input in
-            [yY][eE][sS]|[yY])
-                echo 'Modify Default IP...'
-                sed -i "s/192.168.1.1/10.0.0.2/;s/192.168/10.0/" package/base-files/files/bin/config_generate
-
-                echo -e '***Done***\n'
-                ;;
-
-            [nN][oO]|[nN])
-                ;;
-
-            *)
-                echo "Invalid input..."
-                exit 1
-                ;;
-        esac
-        ;;
-
-    *)
-        echo "Invalid input..."
-        exit 1
-        ;;
-esac
-
-read -r -p "Reset feeds? [Y/n] " input
-
-case $input in
-    [yY][eE][sS]|[yY])
-        echo 'Reset feeds'
-        rm -rf tmp
-        ./scripts/feeds clean
-        ./scripts/feeds update -a && ./scripts/feeds install -a
-        echo -e '***Done***\n'
-
-        if [ -L $CONF_DIR/${INPUT}_defconfig ]; then
-            echo 'restore last config'
-            cp $CONF_DIR/${INPUT}_defconfig .config
-        fi
-        ;;
-
-    *)
-        ;;
-esac
+    esac
+fi
